[REFACTOR]: Add SparkleText component and update import path in Readme, Tailwind config with new colors 🧩
