+++ components/home/Readme.tsx
@@ -3,7 +3,7 @@ import Link from "next/link";
 import { FaGithub, FaLinkedin } from "react-icons/fa";
 import { FaXTwitter } from "react-icons/fa6";
 
-import SparkleText from "../SparkleText";
+import SparkleText from "./SparkleText";
 import useGitHubFollowers from "@/../hooks/useGitHubFollowers";
 
 const Readme = () => {
